style: clear clippy 1.95.0 lint drift across reddb-server
The rust-toolchain.toml bump to 1.95.0 introduced stricter clippy lints that
were masked behind the fmt failure on main. Apply machine-applicable fixes
(`cargo clippy --fix`) plus manual behavior-preserving fixes:
sort_by→sort_by_key, needless_range_loop→enumerate, drop dead
..Default::default(), &mut Vec→&mut slice, if-let over expect-after-is_some,
collapse identical PEEK-count branches, doc-comment list indentation. One
#[allow(clippy::large_enum_variant)] on the query Filter enum (boxing would
ripple through all call sites).

`cargo clippy --locked --all -- -D warnings` and `cargo check --locked --all`
both green.

+++ crates/reddb-server/src/runtime/impl_core.rs
@@ -6479,7 +6479,7 @@ impl RedDBRuntime {
                             crate::storage::schema::Value::UnsignedInteger(t) => {
                                 i64::try_from(*t).ok()
                             }
-                            crate::storage::schema::Value::Integer(t) => Some(*t as i64),
+                            crate::storage::schema::Value::Integer(t) => Some(*t),
                             _ => None,
                         }),
                 };
